tema 3/requirements.txt
flask==3.1.3
gunicorn==23.0.0
google-cloud-storage
google-cloud-translate
google-cloud-datastore
google-cloud-language
google-cloud-texttospeech